wip: execute command line
