+++ crates/components/command-line/src/ast.rs
@@ -1,18 +1,45 @@
+use mizer_command_executor::{CommandImpl, SendableCommand};
+use mizer_fixtures::definition::FixtureFaderControl;
 use crate::parser::*;
 use mizer_fixtures::FixtureId;
 
-pub enum Ast {
-    Delete(Target),
-    Select(Target),
-    Store(Target),
-    Highlight,
-}
-
-pub enum Target {
-    Fixture(FixtureTarget),
-    Sequence(u32),
-    Group(u32),
-}
+// pub enum Ast {
+//     Delete(Target),
+//     Select(Target),
+//     Store(Target),
+//     Highlight,
+//     Clear,
+//     Write(FixtureTarget, FixtureFaderControl, f64),
+// }
+// 
+// pub struct Fixtures;
+// pub struct Sequences;
+// pub struct Groups;
+// 
+// pub struct ActiveSelection;
+// pub struct Single(u32);
+// pub struct Range(u32, u32);
+// 
+// pub struct Delete;
+// pub struct Store;
+// pub struct Select;
+// pub struct Highlight;
+// pub struct Clear;
+// pub struct Write;
+// 
+// trait Executable {
+//     type Source;
+//     type Command;
+//     type Target;
+// }
+// 
+// impl Executable for () {
+//     type Source = ();
+//     type Command = Select;
+//     type Target = (Fixtures, Single);
+// }
+// 
+// struct Command<TSource, TCommand, TTarget>(TSource, TCommand, TTarget);
 
 pub fn parse(tokens: Tokens) -> anyhow::Result<Ast> {
     let mut tokens = tokens.into_iter();
@@ -39,34 +66,20 @@ pub fn parse(tokens: Tokens) -> anyhow::Result<Ast> {
         Token::Action(Action::Select) => {
             let target = match tokens.next().unwrap() {
                 Token::Target((Keyword::Fixture, Selection::Single(id))) => Target::Fixture(FixtureTarget::Single(FixtureId::from(id as u32))),
-                Token::Target((Keyword::Fixture, Selection::Range(start, end))) => Target::Fixture(FixtureTarget::Range((FixtureId::from(start as u32), FixtureId::from(end as u32)))),
+                Token::Target((Keyword::Fixture, Selection::Range(start, end))) => Target::Fixture(FixtureTarget::Range((start as u32, end as u32))),
                 _ => todo!(),
             };
             Ok(Ast::Select(target))
         }
         Token::Action(Action::Highlight) => Ok(Ast::Highlight),
+        Token::Action(Action::Clear) => Ok(Ast::Clear),
+        Token::Operator(Operator::At) => todo!(),
         _ => todo!(),
     }
 }
 
 pub enum FixtureTarget {
     Single(FixtureId),
-    Range((FixtureId, FixtureId)),
-}
-
-impl From<FixtureTarget> for Vec<FixtureId> {
-    fn from(target: FixtureTarget) -> Self {
-        match target {
-            FixtureTarget::Single(id) => vec![id],
-            FixtureTarget::Range((start, end)) => {
-                // TODO: should the select command discard the unknown fixture ids or the command line executor?
-                let mut ids = Vec::new();
-                for id in start..=end {
-                    ids.push(FixtureId::from(id));
-                }
-
-                ids
-            },
-        }
-    }
+    Range((u32, u32)),
+    Selection,
 }
