+++ crates/components/command-line/src/parser/mod.rs
@@ -30,6 +30,7 @@ pub enum Action {
     Delete,
     Off,
     Highlight,
+    Clear,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -57,8 +58,59 @@ pub enum Selection {
 pub fn parse(input: &str) -> anyhow::Result<Tokens> {
     let result = parser::lexer().parse(input);
     println!("{:?}", result);
-    let ast = result.into_output().unwrap();
+    let (output, errors) = result.into_output_errors();
+    let ast = output.ok_or_else(|| anyhow::anyhow!("Failed to parse: {:?}", errors))?;
 
     Ok(ast)
 }
 
+#[cfg(test)]
+mod tests {
+    use test_case::test_case;
+    use super::*;
+
+    #[test_case("@ 20", 20)]
+    #[test_case("@ 100", 100)]
+    #[test_case("at 20", 20)]
+    fn parse_at(input: &str, expected: i64) {
+        let ast = parse(input).unwrap();
+
+        assert_eq!(ast, Tokens(vec![
+            Token::Operator(Operator::At),
+            Token::Number(expected),
+        ]));
+    }
+
+    #[test_case("select fixtures 1", 1)]
+    #[test_case("select fixtures 3", 3)]
+    fn parse_fixture_selection(input: &str, expected: i64) {
+        let ast = parse(input).unwrap();
+
+        assert_eq!(ast, Tokens(vec![
+            Token::Action(Action::Select),
+            Token::Target((Keyword::Fixture, Selection::Single(expected))),
+        ]));
+    }
+
+    #[test_case("select fixtures 1..3", 1, 3)]
+    #[test_case("select fixtures 2 .. 4", 2, 4)]
+    #[test_case("select fixtures 3 thru 5", 3, 5)]
+    fn parse_multiple_fixture_selection(input: &str, from: i64, to: i64) {
+        let ast = parse(input).unwrap();
+
+        assert_eq!(ast, Tokens(vec![
+            Token::Action(Action::Select),
+            Token::Target((Keyword::Fixture, Selection::Range(from, to))),
+        ]));
+    }
+
+    #[test_case("delete sequence 1", 1)]
+    fn parse_delete_sequence(input: &str, expected: i64) {
+        let ast = parse(input).unwrap();
+
+        assert_eq!(ast, Tokens(vec![
+            Token::Action(Action::Delete),
+            Token::Target((Keyword::Sequence, Selection::Single(expected))),
+        ]));
+    }
+}
