+++ docs/antora-playbook.yml
@@ -15,6 +15,8 @@ antora:
     - '@antora/lunr-extension'
     - '@antora/pdf-extension'
 asciidoc:
+  extensions:
+    - asciidoctor-kroki
   attributes:
     page-pagination: ''
     experimental: ''
@@ -24,4 +26,5 @@ asciidoc:
     example-caption: false
     listing-caption: false
     hide-uri-scheme: '@'
+    kroki-fetch-diagram: true
