+++ docs/modules/console/nav.adoc
@@ -1,0 +1,3 @@
+.Command Line
+* xref:console:grammar.adoc[]
+* xref:console:examples.adoc[]
