+++ docs/modules/console/pages/grammar.adoc
@@ -1,0 +1,67 @@
+= Grammar
+
+.Select Fixtures 1 Thru 10
+[graphviz, format=svg]
+....
+include::partial$diagrams/select_fixtures.dot[]
+....
+
+.1 Thru 10 @ 20
+[graphviz, format=svg]
+....
+include::partial$diagrams/write_fixtures.dot[]
+....
+
+.@ Full
+[graphviz, format=svg]
+....
+include::partial$diagrams/write_fixture_selection.dot[]
+....
+
+.Channel Pan @ 30
+[graphviz, format=svg]
+....
+include::partial$diagrams/write_channel.dot[]
+....
+
+.1 Thru 10 Channel Pan @ 30
+[graphviz, format=svg]
+....
+include::partial$diagrams/write_channel_fixtures.dot[]
+....
+
+.Delete Group 1
+[graphviz, format=svg]
+....
+include::partial$diagrams/delete_group.dot[]
+....
+
+.Delete Sequence 1 Thru 10
+[graphviz, format=svg]
+....
+include::partial$diagrams/delete_sequence.dot[]
+....
+
+.Highlight
+[graphviz, format=svg]
+....
+include::partial$diagrams/highlight.dot[]
+....
+
+.Store Group 1
+[graphviz, format=svg]
+....
+include::partial$diagrams/store_group.dot[]
+....
+
+.Store Sequence 1
+[graphviz, format=svg]
+....
+include::partial$diagrams/store_sequence.dot[]
+....
+
+.Select Sequence 1
+[graphviz, format=svg]
+....
+include::partial$diagrams/select_sequence.dot[]
+....
