+++ docs/modules/console/partials/diagrams/delete_group.dot
@@ -1,0 +1,5 @@
+// Delete Group 1
+strict graph {
+    DELETE -- GROUP
+    GROUP -- 1
+}
