+++ docs/modules/console/partials/diagrams/highlight.dot
@@ -1,0 +1,4 @@
+// Highlight
+strict graph {
+    HIGHLIGHT
+}
