+++ docs/modules/console/partials/diagrams/select_sequence.dot
@@ -1,0 +1,5 @@
+// Select Sequence 1
+strict graph {
+    SELECT -- SEQUENCES
+    SEQUENCES -- 1
+}
