+++ docs/modules/console/partials/diagrams/store_group.dot
@@ -1,0 +1,6 @@
+// Store Group 1
+strict graph {
+    STORE -- "ACTIVE SELECTION"
+    STORE -- GROUP
+    GROUP -- 1
+}
