+++ docs/modules/console/partials/diagrams/store_sequence.dot
@@ -1,0 +1,6 @@
+// Store Sequence 1
+strict graph {
+    STORE -- "ACTIVE PROGRAMMER"
+    STORE -- SEQUENCES
+    SEQUENCES -- 1
+}
