+++ docs/modules/console/partials/diagrams/write_fixtures.dot
@@ -1,0 +1,9 @@
+// 1 Thru 10 @ 20
+strict graph {
+    WRITE -- FIXTURES
+    FIXTURES -- "1 Thru 10"
+    WRITE -- CHANNEL
+    CHANNEL -- INTENSITY
+    WRITE -- VALUE
+    VALUE -- 20
+}
