+++ docs/package.json
@@ -14,6 +14,7 @@
     "@antora/cli": "3.1.9",
     "@antora/pdf-extension": "^1.0.0-alpha.9",
     "@antora/site-generator": "3.1.9",
+    "asciidoctor-kroki": "^0.18.1",
     "http-server": "^14.1.1"
   },
   "dependencies": {
